Refactor: split render into multiple files
