+++ src/flowrep/api/tools.py
@@ -21,8 +21,8 @@ Use `run_recipe` to convert recipes into retrospective instance objects with out
 data.
 """
 
-from flowrep.compiler.render import dagdata2python as dagdata2python
-from flowrep.compiler.render import workflow2python as workflow2python
+from flowrep.compiler.source import dagdata2python as dagdata2python
+from flowrep.compiler.source import workflow2python as workflow2python
 from flowrep.converters.python_workflow_definition import flowrep2pwd as flowrep2pwd
 from flowrep.converters.python_workflow_definition import pwd2flowrep as pwd2flowrep
 from flowrep.parsers.atomic_parser import atomic as atomic
